Add files via upload
Preliminary update for the Delve league. Includes support for actual Delve content (Reliquaries, Resonators+Fossils, etc), as well as the initial implementation of minimap icons and item beams, and added a highlight for a handful of T1 explicit mods. I'll likely update this again a few days after the launch depending on whether the new features need further tweaks or not.